refactor: extract assignment detail container
